+++ scraper/src/main/java/GroceryFamily/GroceryDad/scraper/view/barbora/BarboraProductListView.java
@@ -1,4 +1,4 @@
-package GroceryFamily.GroceryDad.scraper.context.barbora;
+package GroceryFamily.GroceryDad.scraper.view.barbora;
 
 import GroceryFamily.GroceryDad.scraper.page.Link;
 import GroceryFamily.GroceryDad.scraper.view.ProductListView;
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
 
-import static GroceryFamily.GroceryDad.scraper.context.barbora.BarboraView.productCode;
+import static GroceryFamily.GroceryDad.scraper.view.barbora.BarboraView.productCode;
 
 @SuperBuilder
 class BarboraProductListView extends View implements ProductListView {
