+++ scraper/src/main/java/GroceryFamily/GroceryDad/scraper/view/barbora/BarboraViewFactory.java
@@ -1,18 +1,18 @@
-package GroceryFamily.GroceryDad.scraper.context.prisma;
+package GroceryFamily.GroceryDad.scraper.view.barbora;
 
 import GroceryFamily.GroceryDad.scraper.page.Source;
 import GroceryFamily.GroceryDad.scraper.view.*;
 import org.jsoup.nodes.Document;
 
-public class PrismaViewFactory implements ViewFactory {
+public class BarboraViewFactory implements ViewFactory {
     @Override
     public LiveView liveView() {
-        return PrismaLiveView.INSTANCE;
+        return BarboraLiveView.INSTANCE;
     }
 
     @Override
     public CategoryView categoryView(Document document, Source selected) {
-        return PrismaCategoryView
+        return BarboraCategoryView
                 .builder()
                 .document(document)
                 .selected(selected)
@@ -21,7 +21,7 @@ public class PrismaViewFactory implements ViewFactory {
 
     @Override
     public ProductListView productListView(Document document, Source selected) {
-        return PrismaProductListView
+        return BarboraProductListView
                 .builder()
                 .document(document)
                 .selected(selected)
@@ -30,7 +30,7 @@ public class PrismaViewFactory implements ViewFactory {
 
     @Override
     public ProductView productView(Document document, Source selected) {
-        return PrismaProductView
+        return BarboraProductView
                 .builder()
                 .document(document)
                 .selected(selected)
