+++ scraper/src/main/java/GroceryFamily/GroceryDad/scraper/view/prisma/PrismaProductView.java
@@ -1,4 +1,4 @@
-package GroceryFamily.GroceryDad.scraper.context.rimi;
+package GroceryFamily.GroceryDad.scraper.view.prisma;
 
 import GroceryFamily.GroceryDad.scraper.view.ProductView;
 import GroceryFamily.GroceryDad.scraper.view.View;
@@ -6,19 +6,20 @@ import GroceryFamily.GroceryElders.domain.Namespace;
 import GroceryFamily.GroceryElders.domain.Product;
 import lombok.experimental.SuperBuilder;
 
-import static GroceryFamily.GroceryDad.scraper.context.rimi.RimiView.productCode;
+import static GroceryFamily.GroceryDad.scraper.view.prisma.PrismaView.productCode;
 
 @SuperBuilder
-class RimiProductView extends View implements ProductView {
+class PrismaProductView extends View implements ProductView {
     @Override
     public Product product() {
         return Product
                 .builder()
-                .namespace(Namespace.RIMI)
+                .namespace(Namespace.PRISMA)
                 .code(productCode(selected.url))
-                .name(document.select("h3[class=name]").text())
+                .name(document.select("#product-name").text())
                 .url(selected.url)
                 // todo: set prices and categories
                 .build();
     }
+
 }
