+++ scraper/src/main/java/GroceryFamily/GroceryDad/scraper/view/rimi/RimiProductView.java
@@ -1,4 +1,4 @@
-package GroceryFamily.GroceryDad.scraper.context.prisma;
+package GroceryFamily.GroceryDad.scraper.view.rimi;
 
 import GroceryFamily.GroceryDad.scraper.view.ProductView;
 import GroceryFamily.GroceryDad.scraper.view.View;
@@ -6,20 +6,17 @@ import GroceryFamily.GroceryElders.domain.Namespace;
 import GroceryFamily.GroceryElders.domain.Product;
 import lombok.experimental.SuperBuilder;
 
-import static GroceryFamily.GroceryDad.scraper.context.prisma.PrismaView.productCode;
-
 @SuperBuilder
-class PrismaProductView extends View implements ProductView {
+class RimiProductView extends View implements ProductView {
     @Override
     public Product product() {
         return Product
                 .builder()
-                .namespace(Namespace.PRISMA)
-                .code(productCode(selected.url))
-                .name(document.select("#product-name").text())
+                .namespace(Namespace.RIMI)
+                .code(RimiView.productCode(selected.url))
+                .name(document.select("h3[class=name]").text())
                 .url(selected.url)
                 // todo: set prices and categories
                 .build();
     }
-
 }
