+++ scraper/src/main/java/GroceryFamily/GroceryDad/scraper/view/rimi/RimiView.java
@@ -1,4 +1,4 @@
-package GroceryFamily.GroceryDad.scraper.context.rimi;
+package GroceryFamily.GroceryDad.scraper.view.rimi;
 
 import GroceryFamily.GroceryDad.scraper.page.Path;
